+++ base/complex.jl
@@ -329,7 +329,7 @@ log2(z::Complex) = log(z)/oftype(real(z),0.6931471805599453)
 function exp(z::Complex)
     zr, zi = reim(z)
     if isnan(zr)
-        Complex(zr, zi==zero(zi) ? zi : zr)
+        Complex(zr, zi==0 ? zi : zr)
     elseif !isfinite(zi)
         if zr == inf(zr)
             Complex(-zr, nan(zr))
@@ -351,7 +351,7 @@ end
 function expm1(z::Complex)
     zr,zi = reim(z)
     if isnan(zr)
-        Complex(zr, zi==zero(zi) ? zi : zr)
+        Complex(zr, zi==0 ? zi : zr)
     elseif !isfinite(zi)
         if zr == inf(zr)
             Complex(-zr, nan(zr))
@@ -362,13 +362,31 @@ function expm1(z::Complex)
         end
     else
         erm1 = expm1(zr)        
-        if zi == zero(zi)
+        if zi == 0
             Complex(erm1, zi)
         else            
             er = erm1+one(erm1)
             wr = isfinite(er) ? erm1 - 2.0*er*(sin(0.5*zi))^2 : er*cos(zi)
             Complex(wr, er*sin(zi))
         end
+    end
+end
+
+function log1p{T}(z::Complex{T})
+    zr,zi = reim(z)
+    if isfinite(zr)
+        isinf(zi) && return log(z)
+        # This is based on a well-known trick for log1p of real z,
+        # allegedly due to Kahan, only modified to handle real(u) <= 0
+        # differently to avoid inaccuracy near z==-2 and for correct branch cut
+        u = float(one(T)) + z
+        u == 1 ? convert(typeof(u), z) : real(u) <= 0 ? log(u) : log(u)*z/(u-1)
+    elseif isnan(zr)
+        Complex(zr, zr)
+    elseif isfinite(zi)
+        Complex(inf(T), copysign(zr > 0 ? zero(T) : convert(T, pi), zi))
+    else
+        Complex(inf(T), nan(T))
     end
 end